Add tools for analysing the code and move the sources in src/
