+++ package.json
@@ -1,0 +1,41 @@
+{
+    "name": "random-agent-spoofer",
+    "version": "0.1.0",
+    "main": "index.js",
+    "scripts": {
+        "start": "gulp"
+    },
+    "repository": {
+        "url": "https://github.com/Mylainos/random-agent-spoofer.git",
+        "type": "git"
+    },
+    "author": "Hugo Thiessard <hugo.thiessard@opmbx.org>",
+    "license": "GPL-3.0+",
+    "devDependencies": {
+        "babel-preset-latest": "^6.16.0",
+        "del": "^2.2.2",
+        "eslint": "^3.11.1",
+        "eslint-config-airbnb": "^13.0.0",
+        "eslint-plugin-import": "^2.2.0",
+        "eslint-plugin-jsx-a11y": "2.2.3",
+        "eslint-plugin-react": "^6.7.1",
+        "gulp": "^3.9.1",
+        "gulp-babel": "^6.1.2",
+        "gulp-eslint": "^3.0.1"
+    },
+    "babel": {
+        "presets": [
+            "latest"
+        ]
+    },
+    "eslintConfig": {
+        "extends": "airbnb",
+        "plugins": [
+            "import"
+        ],
+        "env": {
+            "browser": true,
+            "webextensions": true
+        }
+    }
+}
