+++ src/data/images/.directory
