+++ src/data/images/whitelist_enabled.png
